chore: move react utils to plugin react
