+++ packages/compat/plugin-swc/package.json
@@ -41,6 +41,7 @@
     "@babel/preset-react": "^7.22.15",
     "@babel/preset-typescript": "^7.23.2",
     "@modern-js/swc-plugins": "0.6.5",
+    "@rsbuild/plugin-react": "workspace:*",
     "@rsbuild/shared": "workspace:*",
     "@swc/helpers": "0.5.3",
     "core-js": "~3.32.2",
