+++ packages/shared/package.json
@@ -101,14 +101,12 @@
     "caniuse-lite": "^1.0.30001559",
     "line-diff": "2.1.1",
     "lodash": "^4.17.21",
-    "postcss": "8.4.31",
-    "semver": "^7.5.4"
+    "postcss": "8.4.31"
   },
   "devDependencies": {
     "@rspack/core": "0.3.14",
     "@types/lodash": "^4.14.200",
     "@types/node": "^16",
-    "@types/semver": "^7.5.4",
     "html-webpack-plugin": "npm:html-rspack-plugin@5.5.5",
     "http-proxy-middleware": "^2.0.1",
     "terser": "5.19.2",
